fix: revert sepalate logics
