+++ app/components/InstanceCard.vue
@@ -93,13 +93,17 @@
 </template>
 
 <script setup lang="ts">
-import { useFormat } from '~/composables/useFormat';
-
 const props = defineProps<{
   instance: Instance;
 }>();
 
-const { formatNumber } = useFormat();
+const formatNumber = (num: number | undefined | null) => {
+  if (num == null) return '-';
+  if (num >= 100000) {
+    return new Intl.NumberFormat('ja-JP', { notation: 'compact', maximumFractionDigits: 1 }).format(num);
+  }
+  return new Intl.NumberFormat('ja-JP').format(num);
+};
 
 const fetchedIcon = ref<string | null>(null);
 const fetchedBanner = ref<string | null>(null);
