+++ app/components/SiteHero.vue
@@ -88,8 +88,6 @@
 </template>
 
 <script setup lang="ts">
-import { useFormat } from '~/composables/useFormat';
-
 defineProps<{
   stats?: {
     counts: {
@@ -108,7 +106,13 @@ defineProps<{
   };
 }>();
 
-const { formatNumber } = useFormat();
+const formatNumber = (num: number | undefined | null) => {
+  if (num == null) return '-';
+  if (num >= 100000) {
+    return new Intl.NumberFormat('ja-JP', { notation: 'compact', maximumFractionDigits: 1 }).format(num);
+  }
+  return new Intl.NumberFormat('ja-JP').format(num);
+};
 </script>
 
 <style scoped>
